+++ src/main/kotlin/dev/thynanami/idea/typst/configuration/BinarySource.kt
@@ -1,6 +1,6 @@
 package dev.thynanami.idea.typst.configuration
 
 enum class BinarySource {
-  USE_AUTOMATIC_DOWNLOAD,
+  USE_BUNDLED_BINARY,
   USE_CUSTOM_BINARY
 }