+++ src/main/kotlin/dev/thynanami/idea/typst/languageserver/TypstLspServerSupportProvider.kt
@@ -2,12 +2,8 @@ package dev.thynanami.idea.typst.languageserver
 
 import dev.thynanami.idea.typst.TypstIcons
 import dev.thynanami.idea.typst.configuration.SettingsConfigurable
-import dev.thynanami.idea.typst.languageserver.downloader.Filesystem
-import dev.thynanami.idea.typst.languageserver.downloader.TinymistDownloadScheduler
-import dev.thynanami.idea.typst.languageserver.downloader.TinymistDownloader
-import dev.thynanami.idea.typst.languageserver.locations.TinymistLocationResolver
+import dev.thynanami.idea.typst.languageserver.locations.TinymistBinary
 import dev.thynanami.idea.typst.languageserver.locations.isSupportedTypstFileType
-import com.intellij.openapi.diagnostic.logger
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.util.NlsActions
 import com.intellij.openapi.vfs.VirtualFile
@@ -17,15 +13,6 @@ import com.intellij.platform.lsp.api.lsWidget.LspClientWidgetItem
 
 
 class TypstLspServerSupportProvider : LspIntegrationProvider {
-  private val downloadScheduler by lazy {
-    TinymistDownloadScheduler(
-      TinymistLocationResolver(),
-      TinymistDownloader(),
-      Filesystem(),
-      TypstLanguageServerManager()
-    )
-  }
-
   override fun fileOpened(
     project: Project,
     file: VirtualFile,
@@ -35,13 +22,15 @@ class TypstLspServerSupportProvider : LspIntegrationProvider {
       return
     }
 
-    TypstManager(downloadScheduler, project, clientStarter).startIfRequired()
+    clientStarter.ensureClientStarted(
+      TinymistLanguageServerDescriptor(TinymistBinary.resolve(), project)
+    )
   }
 
   override fun createWidgetItem(
     lspClient: LspClient,
     currentFile: VirtualFile?
-  ): LspClientWidgetItem? {
+  ): LspClientWidgetItem {
     return object : LspClientWidgetItem(
       lspClient,
       currentFile,