+++ src/main/kotlin/dev/thynanami/idea/typst/languageserver/locations/LocationResolver.kt
@@ -1,9 +1,0 @@
-package dev.thynanami.idea.typst.languageserver.locations
-
-import java.net.URI
-import java.nio.file.Path
-
-interface LocationResolver {
-  fun downloadUrl(): URI
-  fun binaryPath(): Path
-}